JSON schema draft
Fixes #30

This JSON schema describes the structure of an AI BOM document, based on the attributes you provided. It defines which fields are required and which are optional, as well as the expected data types for each field. You can use this schema to validate any AI BOM documents to ensure they meet the specification you've outlined. Later we should have a common name space that is independent of a specific SBOM standard (allowing any standard to make a reference to it).